+++ mycodo/inputs/mh_z16.py
@@ -22,8 +22,7 @@
 import logging
 import time
 
-import fasteners
-import struct
+import locket
 
 from .base_input import AbstractInput
 from .sensorutils import is_device
@@ -108,6 +107,7 @@ class MHZ16Sensor(AbstractInput):
 
     def get_measurement(self):
         """ Gets the MH-Z16's CO2 concentration in ppmv via UART"""
+        lock_acquired = False
         self._co2 = None
         co2 = None
 
@@ -116,17 +116,13 @@ class MHZ16Sensor(AbstractInput):
             if not self.serial_device:  # Don't measure if device isn't validated
                 return None
 
-            # Acquire lock on MHZ16 to ensure more than one read isn't
-            # being attempted at once on the same interface
-            lock = fasteners.InterProcessLock(self.mhz16_lock_file)
-            lock_acquired = False
-
-            for _ in range(600):
-                lock_acquired = lock.acquire(blocking=False)
-                if lock_acquired:
-                    break
-                else:
-                    time.sleep(0.1)
+            # Set up lock
+            lock = locket.lock_file(self.mhz16_lock_file, timeout=60)
+            try:
+                lock.acquire()
+                lock_acquired = True
+            except:
+                self.logger.error("Could not acquire lock.")
 
             if lock_acquired:
                 self.ser.flushInput()
@@ -139,8 +135,6 @@ class MHZ16Sensor(AbstractInput):
                     low_level = resp[3]
                     co2 = high_level * 256 + low_level
                 lock.release()
-            else:
-                self.logger.error("Could not acquire MHZ16 lock")
 
         elif self.interface == 'I2C':
             self.write_register(self.FCR, 0x07)